requirements.txt
anthropic>=0.40.0
pyautogui>=0.9.54
mss>=9.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
